LayoutTests/inspector/console/console-external-array.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function onload()
{
    console.log(new Int8Array(10));
    console.log(new Int16Array(10));
    console.log(new Int32Array(10));
    console.log(new Uint8Array(10));
    console.log(new Uint16Array(10));
    console.log(new Uint32Array(10));
    console.log(new Float32Array(10));
    console.log(new Float64Array(10));

    console.dir(new Int8Array(10));
    console.dir(new Int16Array(10));
    console.dir(new Int32Array(10));
    console.dir(new Uint8Array(10));
    console.dir(new Uint16Array(10));
    console.dir(new Uint32Array(10));
    console.dir(new Float32Array(10));
    console.dir(new Float64Array(10));

    runTest();
}

function test()
{
    InspectorTest.dumpConsoleMessages();
    InspectorTest.completeTest();
}

</script>
</head>

<body onload="onload()">
<p>
Tests that console logging detects external arrays as arrays.
</p>

</body>
</html>
